# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 우리 언제? API
  version: 0.1.0
  description: |
    기간 기반 모임 시간 조율 서비스 "우리 언제?"의 API 계약(단일 출처, SSOT).

    설계-퍼스트로 이 명세를 먼저 합의하고, 프론트엔드는 여기서 타입을 생성하며
    백엔드(NestJS)는 이 명세에 맞춰 구현한다.

    - 도메인 규칙: docs/domain.md / plan/whenwe_plan_v0.3.pdf
    - 모든 응답은 `{ success, data, error }` 봉투로 감싼다 (@whenwe/types 의 ApiResponse 와 일치).
    - 인증: 모임장 = JWT Bearer, 참여자 응답 수정 = X-Participant-Edit-Token 헤더,
      초대 접근 = inviteToken 경로 파라미터.
    - 날짜·시간은 ISO 8601(date-time 은 타임존 오프셋 포함, 예: +09:00).
servers:
  - url: http://localhost:3001
    description: 로컬 개발
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Auth]
      operationId: signup
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: 가입 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AuthUser' }
        '409':
          $ref: '#/components/responses/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 로그인 (JWT 발급)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LoginResult' }
        '401':
          $ref: '#/components/responses/Error'
  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 로그아웃
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Empty' }
        '401':
          $ref: '#/components/responses/Error'
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: 현재 로그인 사용자 조회
      responses:
        '200':
          description: 사용자 정보
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AuthUser' }
        '401':
          $ref: '#/components/responses/Error'

  # ──────────────── Meetings ────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 모임장(로그인 사용자) 인증
    participantToken:
      type: apiKey
      in: header
      name: X-Participant-Edit-Token
      description: 비회원 참여자의 응답 수정 토큰
...
  responses:
    Error:
      description: 에러 (공통 봉투)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ApiSuccess:
      type: object
      required: [success, data, error]
      properties:
        success: { const: true }
        error: { type: 'null' }
        data: {}
    ErrorResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { const: false }
        data: { type: 'null' }
        error: { $ref: '#/components/schemas/ApiError' }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
    Empty:
      type: object
      additionalProperties: false

    # ── enums (값은 @whenwe/types 와 일치) ──
...
    ErrorCode:
      type: string
      enum:
        - MEETING_NOT_FOUND
        - INVITE_TOKEN_INVALID
        - INVITE_TOKEN_EXPIRED
        - RESPONSE_DEADLINE_PASSED
        - MEETING_ALREADY_CONFIRMED
        - PARTICIPANT_EDIT_TOKEN_INVALID
        - FORBIDDEN_MEETING_OWNER_ONLY
        - RECOMMENDATION_NOT_READY
        - EMAIL_ALREADY_EXISTS
        - INVALID_CREDENTIALS
        - UNAUTHENTICATED

    # ── Auth ──
    SignupRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        nickname: { type: string, minLength: 1, maxLength: 30 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        nickname: { type: string }
    LoginResult:
      type: object
      properties:
        accessToken: { type: string }
        user: { $ref: '#/components/schemas/AuthUser' }

    # ── Meetings ──
...
security:
  - bearerAuth: []